MGMT-3510 Move static ip script constant to a separate file
Importing `ignition` into `isoeditor` was causing an import cycle
and nothing in `ignition` uses the script.

The idea here is that any package can import `constants` without
worrying about a cycle because it will never import any other package.
